Make code neater.
