+++ travel-planner/travel-planner-ui/src/Components/Maps/Maps.css
@@ -97,7 +97,7 @@
 	height: 0;
 	position: absolute;
 	/* The max width of the info window. */
-	width: 12vw;
+	width: 15vw;
 }
 
 .info-window-top-container {
