+++ travel-planner/travel-planner-ui/src/Components/Marker/Marker.jsx
@@ -1,0 +1,127 @@
+import { useState, useEffect } from 'react'
+import {render} from 'react-dom'
+
+export default function Marker(props) {
+	class InfoWindow extends google.maps.OverlayView {
+		constructor(position, adjustTitle, setAdjustTitle, curTitle, setCurTitle) {
+			super();
+
+			this.position = position;
+			this.new = true;
+
+			// Create the content displayed when user opens the infoWindow
+			const content = (
+				<div className="info-window-content popup-bubble" id={`${curTitle}`}>
+					<div className="info-window-top-container">
+						{adjustTitle ? (
+							<input
+								className="info-window-title"
+								onKeyDown={(event) => {
+									if (event.key === "Enter") {
+										setAdjustTitle(false);
+									}
+								}}
+								onChange={(event) => setCurTitle(event.currentTarget.value)}
+								value={curTitle}
+							/>
+						) : (
+							<h2
+								onDoubleClick={() => setAdjustTitle(true)}
+								className="info-window-title"
+							>
+								{curTitle}
+							</h2>
+						)}
+						<button
+							onClick={() => this.onRemove(this)}
+							className="close-info-window-button"
+						>
+							x
+						</button>
+					</div>
+					<p>Address: from search term or inverse geocode?</p>
+					<div className="info-window-note">
+						<button> Open Note </button>
+					</div>
+				</div>
+			);
+
+			// Create the little arrow beneath the content.
+			const bubbleAnchor = <div className="popup-bubble-anchor">{content}</div>;
+
+			// Hold the content in a HTML node container to work with the Google Maps Javascript API
+			this.containerDiv = document.createElement("div");
+			this.containerDiv.classList.add("popup-container");
+			this.containerDiv.style.display = "none";
+
+			render(bubbleAnchor, this.containerDiv);
+		}
+
+		onAdd() {
+			// This function is automatically called when the popup is added to the map
+			if (this.new) {
+				this.new = false;
+			} else {
+				this.getPanes().floatPane.appendChild(this.containerDiv);
+			}
+		}
+
+		onRemove(newThis = this) {
+			// Automatically called when popup is removed from the map
+
+			if (newThis.containerDiv.parentElement) {
+				newThis.containerDiv.parentElement.removeChild(newThis.containerDiv);
+			}
+		}
+
+		draw() {
+			// Called each frame when the popup needs to draw itself
+
+			const newDivPosition = this.getProjection().fromLatLngToDivPixel(
+				this.position
+			);
+
+			// Hide the popup when it is out of view
+			const display =
+				Math.abs(newDivPosition.x) < 4000 && Math.abs(newDivPosition.y) < 4000
+					? "block"
+					: "none";
+
+			if (display === "block") {
+				this.containerDiv.style.left = newDivPosition.x + "px";
+				this.containerDiv.style.top = newDivPosition.y + "px";
+			}
+
+			if (this.containerDiv.style.display !== display) {
+				this.containerDiv.style.display = display;
+			}
+		}
+	}
+
+	const [adjustTitle, setAdjustTitle] = useState(false);
+	const [curTitle, setCurTitle] = useState(props.title);
+
+	useEffect(() => {
+		const marker = new google.maps.Marker(props);
+
+		const popup = new InfoWindow(
+			props.position,
+			adjustTitle,
+			setAdjustTitle,
+			curTitle,
+			setCurTitle
+		);
+		popup.setMap(props.map);
+		popup.onRemove();
+
+		marker.addListener("click", () => {
+			popup.onAdd();
+		});
+
+		return () => {
+			marker.setMap(null);
+		};
+	}, [props]);
+
+	return null;
+}
